Update comment endpoints
