+++ controllers/comments.ts
@@ -1,79 +1,8 @@
 import { formatError } from "../library/errors";
 import prisma from "../prisma/client";
 import asyncHandler from "express-async-handler";
-
-const getAllComments = async (contentIds: string[]) => {
-  const [userComments, nostrComments] = await Promise.all([
-    prisma.comment.findMany({
-      where: {
-        AND: [
-          {
-            OR: contentIds.map((id) => ({
-              contentId: id,
-            })),
-          },
-          { isNostr: false },
-        ],
-      },
-    }),
-    prisma.comment.findMany({
-      where: {
-        AND: [
-          {
-            OR: contentIds.map((id) => ({
-              contentId: id,
-            })),
-          },
-          { isNostr: true },
-        ],
-      },
-    }),
-  ]);
-
-  const commentsWithUserInfo = await Promise.all(
-    userComments.map(async (comment) => {
-      if (comment.userId === "keysend") {
-        return comment;
-      }
-
-      const user = await prisma.user.findUnique({
-        where: { id: comment.userId },
-      });
-
-      return {
-        ...comment,
-        name: user.name,
-        commenterProfileUrl: user.profileUrl,
-        commenterArtworkUrl: user.artworkUrl,
-      };
-    })
-  );
-
-  const sortedComments = [...commentsWithUserInfo, ...nostrComments].sort(
-    (a, b) => (a.createdAt < b.createdAt ? 1 : -1)
-  );
-
-  const commentsWithSatAmount = await Promise.all(
-    sortedComments.map(async (comment) => {
-      if (comment.txId) {
-        const preamp = await prisma.preamp.findUnique({
-          where: { txId: comment.txId },
-        });
-
-        return {
-          ...comment,
-          commentMsatSum: preamp.msatAmount,
-        };
-      } else {
-        return {
-          ...comment,
-        };
-      }
-    })
-  );
-
-  return commentsWithSatAmount;
-};
+import { getAllComments } from "../library/comments";
+import db from "../library/db";
 
 const get_comments = asyncHandler(async (req, res, next) => {
   const { id: contentId } = req.params;
@@ -111,7 +40,8 @@ const get_podcast_comments = asyncHandler(async (req, res, next) => {
   res.json({ success: true, data: combinedAndSortedComments });
 });
 
-// looks up all album ids for an artist, then all the track ids for each album, and then gets all comments for those tracks
+// looks up all album ids for an artist, then all the track ids for each album,
+// and then gets all comments for those tracks
 const get_artist_comments = asyncHandler(async (req, res, next) => {
   const { id: artistId } = req.params;
   if (!artistId) {
@@ -132,11 +62,38 @@ const get_artist_comments = asyncHandler(async (req, res, next) => {
     },
   });
 
-  const combinedAndSortedComments = await getAllComments(
-    tracks.map(({ id }) => id)
-  );
+  const comments = await getAllComments(tracks.map(({ id }) => id));
 
-  res.json({ success: true, data: combinedAndSortedComments });
+  const commentsLegacy = await db
+    .knex("comment")
+    .leftOuterJoin("user", "comment.user_id", "=", "user.id")
+    .join("amp", "comment.amp_id", "=", "amp.id")
+    .join("track", "track.id", "=", "amp.track_id")
+    .join("artist", "artist.id", "=", "track.artist_id")
+    .select(
+      "comment.id as id",
+      "track.id as trackId",
+      "is_nostr as isNostr",
+      "amp.tx_id as txId"
+    )
+    .min("track.title as title")
+    .min("artist.user_id as ownerId")
+    .min("comment.content as content")
+    .min("comment.created_at as createdAt")
+    .min("amp.msat_amount as commentMsatSum")
+    .min("comment.user_id as userId")
+    .min("user.name as name")
+    .min("user.profile_url as commenterProfileUrl")
+    .min("user.artwork_url as commenterArtworkUrl")
+    .where("artist.id", "=", artistId)
+    .andWhere("amp.comment", "=", true)
+    .andWhere("track.deleted", "=", false)
+    .whereNull("comment.parent_id")
+    .groupBy("comment.id", "track.id", "amp.tx_id")
+    .orderBy("createdAt", "desc");
+
+  // TODO: Pagination
+  res.json({ success: true, data: [...comments, ...commentsLegacy] });
 });
 
 export default {
